Add in new edge-case for Android
