+++ test/fixtures/files/Main.java
@@ -1,35 +1,8 @@
 package com.some.example;
 
-import com.facebook.react.ReactPackage;
-import com.facebook.react.bridge.JavaScriptModule;
-import com.facebook.react.bridge.NativeModule;
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.uimanager.ViewManager;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-public class SomeExamplePackage implements ReactPackage {
-
-  public SomeExamplePackage() {}
-
-  @Override
-  public List<NativeModule> createNativeModules(
-      ReactApplicationContext reactContext) {
-    List<NativeModule> modules = new ArrayList<>();
-    modules.add(new SomeExampleModule(reactContext));
-    return modules;
-  }
-
-  @Override
-  public List<Class<? extends JavaScriptModule>> createJSModules() {
-    return Collections.emptyList();
-  }
-
-  @Override
-  public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-    return Collections.emptyList();
-  }
-}
+public class SomeExamplePackage {}
