+++ test/fixtures/projects/flatStructure.js
@@ -3,11 +3,15 @@ const path = require('path');
 const android = require('./android');
 
 module.exports = {
-  android,
+  android: android.valid,
   node_modules: {
     'react-native-vector-icons': {
-      'package.json': fs.readFileSync(path.join(__dirname, '../files/vector-icons.json')),
-      android,
+      'package.json': fs.readFileSync(path.join(__dirname, '../files/dependency.json')),
+      android: android.valid,
+    },
+    'rnpm-android-no-package': {
+      'package.json': fs.readFileSync(path.join(__dirname, '../files/dependency.json')),
+      android: android.noPackage,
     },
   },
 };
